+++ packages/plugin/src/Services/SettingsService.php
@@ -441,6 +441,7 @@ class SettingsService extends BaseService
 
         $nav = [
             'general' => ['title' => Freeform::t('General Settings')],
+            'form-builder' => ['title' => Freeform::t('Form Builder Settings')],
             'formatting-templates' => ['title' => Freeform::t('Formatting Templates')],
             'email-templates' => ['title' => Freeform::t('Email Templates')],
             'statuses' => ['title' => Freeform::t('Statuses')],